Created all models needed for the database in models.py, and setout possible methods that we may use. This is the first full version however will be updated shortly to normalise the database further and also due to figuring out what data we will be importing from APIs.
Further: Improved all models, added Metadata so that in the Django admin view all models have readable names, and added str methods so that when a model instance is printed, it prints something useful. I pushed all models to the django admin so that they can be seen on the admin pages. I re-sorted the html templates for clarity.